requirements.txt
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.1
fastapi==0.115.11
faiss-cpu==1.10.0
filelock==3.17.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
langchain==0.3.20
langchain-core==0.3.41
langchain-community==0.3.11
langchain-huggingface==0.1.1
MarkupSafe==3.0.2
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentence-transformers==2.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
